api/requirements.txt
fastapi
uvicorn
onnxruntime-gpu
numpy
pymongo
gradio
